requirements.txt
rerun-sdk>=0.16.0
numpy
scipy
h5py
opencv-python
trimesh
urdf_parser_py
pycollada
tensorflow
tensorflow_datasets
gcsfs